+++ packages/base64url/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @phantom/base64url
 
+## 1.0.0-beta.3
+
 ## 1.0.0-beta.2
 
 ## 1.0.0-beta.1