+++ packages/client/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @phantom/client
 
+## 1.0.0-beta.3
+
+### Patch Changes
+
+- 1935986: Fix hooks interfaces and add signAllTransactions
+- Updated dependencies [a290d0e]
+  - @phantom/constants@1.0.0-beta.3
+  - @phantom/base64url@1.0.0-beta.3
+  - @phantom/sdk-types@1.0.0-beta.3
+  - @phantom/crypto@1.0.0-beta.3
+  - @phantom/api-key-stamper@1.0.0-beta.3
+
 ## 1.0.0-beta.2
 
 ### Patch Changes